--- apps/ios/Flashcards/Flashcards/Review/Notifications/ReviewNotificationsSupport.swift
@@ -1,1211 +1,0 @@
-import Foundation
-import UIKit
-import UserNotifications
-
-let reviewNotificationPermissionPromptThreshold: Int = 6
-let defaultDailyReminderHour: Int = 10
-let defaultDailyReminderMinute: Int = 0
-let dailyReminderSchedulingHorizonDays: Int = 7
-let appNotificationPendingRequestsLimit: Int = 64
-let notificationSchedulingDelayedReadbackNanoseconds: UInt64 = 350_000_000
-let defaultInactivityReminderWindowEndHour: Int = 19
-let defaultInactivityReminderWindowEndMinute: Int = 0
-let reviewNotificationFallbackBodyText: String = String(
-    localized: "review_notification.fallback_body",
-    table: "Foundation",
-    comment: "Fallback review notification body text"
-)
-
-let reviewNotificationsSettingsUserDefaultsKeyPrefix: String = "review-notifications-settings::"
-let reviewNotificationPromptStateUserDefaultsKey: String = "review-notification-prompt-state"
-let reviewNotificationSuccessfulReviewCountUserDefaultsKey: String = "review-notification-successful-review-count"
-let reviewNotificationScheduledPayloadsUserDefaultsKeyPrefix: String = "review-notification-scheduled-payloads::"
-let reviewNotificationLastActiveAtUserDefaultsKey: String = "review-notification-last-active-at"
-let appNotificationTapTypeUserInfoKey: String = "appNotificationTapType"
-let pendingAppNotificationTapUserDefaultsKey: String = "pending-app-notification-tap"
-let pendingAppNotificationTapSchemaVersion: Int = 1
-
-enum ReviewNotificationMode: String, Codable, CaseIterable, Identifiable, Hashable, Sendable {
-    case daily
-    case inactivity
-
-    var id: String {
-        self.rawValue
-    }
-
-    var title: String {
-        switch self {
-        case .daily:
-            return String(
-                localized: "review_notification.mode.daily",
-                table: "Foundation",
-                comment: "Daily review notification mode title"
-            )
-        case .inactivity:
-            return String(
-                localized: "review_notification.mode.inactivity",
-                table: "Foundation",
-                comment: "Cards review notification mode title"
-            )
-        }
-    }
-}
-
-enum ReviewNotificationsReconcileTrigger: Hashable, Sendable {
-    case appActive
-    case appBackground
-    case settingsChanged
-    case permissionChanged
-    case reviewRecorded
-    case filterChanged
-    case workspaceChanged
-
-    /// `appActive` clears delivered reminders because the user has returned to the
-    /// app and the old reminders have served their purpose. `reviewRecorded` also
-    /// clears them because the moment a card is reviewed, any "review reminder"
-    /// notification (and the icon badge it carries) is no longer relevant.
-    var shouldClearDeliveredReviewNotifications: Bool {
-        switch self {
-        case .appActive, .reviewRecorded:
-            return true
-        case .appBackground, .settingsChanged, .permissionChanged, .filterChanged, .workspaceChanged:
-            return false
-        }
-    }
-
-    var diagnosticValue: String {
-        switch self {
-        case .appActive:
-            return "app_active"
-        case .appBackground:
-            return "app_background"
-        case .settingsChanged:
-            return "settings_changed"
-        case .permissionChanged:
-            return "permission_changed"
-        case .reviewRecorded:
-            return "review_recorded"
-        case .filterChanged:
-            return "filter_changed"
-        case .workspaceChanged:
-            return "workspace_changed"
-        }
-    }
-}
-
-struct DailyReviewNotificationsSettings: Codable, Hashable, Sendable {
-    let hour: Int
-    let minute: Int
-}
-
-struct InactivityReviewNotificationsSettings: Codable, Hashable, Sendable {
-    let windowStartHour: Int
-    let windowStartMinute: Int
-    let windowEndHour: Int
-    let windowEndMinute: Int
-    let idleMinutes: Int
-}
-
-struct ReviewNotificationsSettings: Hashable, Sendable {
-    let isEnabled: Bool
-    let selectedMode: ReviewNotificationMode
-    let daily: DailyReviewNotificationsSettings
-    let inactivity: InactivityReviewNotificationsSettings
-    let showAppIconBadge: Bool
-}
-
-extension ReviewNotificationsSettings: Codable {
-    private enum CodingKeys: String, CodingKey {
-        case isEnabled
-        case selectedMode
-        case daily
-        case inactivity
-        case showAppIconBadge
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.isEnabled = try container.decode(Bool.self, forKey: .isEnabled)
-        self.selectedMode = try container.decode(ReviewNotificationMode.self, forKey: .selectedMode)
-        self.daily = try container.decode(DailyReviewNotificationsSettings.self, forKey: .daily)
-        self.inactivity = try container.decode(InactivityReviewNotificationsSettings.self, forKey: .inactivity)
-        // Missing key defaults to ON so existing users get the badge automatically.
-        self.showAppIconBadge = try container.decodeIfPresent(Bool.self, forKey: .showAppIconBadge) ?? true
-    }
-
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        try container.encode(self.isEnabled, forKey: .isEnabled)
-        try container.encode(self.selectedMode, forKey: .selectedMode)
-        try container.encode(self.daily, forKey: .daily)
-        try container.encode(self.inactivity, forKey: .inactivity)
-        try container.encode(self.showAppIconBadge, forKey: .showAppIconBadge)
-    }
-}
-
-struct NotificationPermissionPromptState: Codable, Hashable, Sendable {
-    let hasShownPrePrompt: Bool
-    let hasRequestedSystemPermission: Bool
-    let hasDismissedPrePrompt: Bool
-}
-
-struct ScheduledReviewNotificationPayload: Hashable, Sendable, Identifiable {
-    let workspaceId: String
-    let reviewFilter: PersistedReviewFilter
-    let content: ScheduledReviewNotificationPayloadContent
-    let scheduledAtMillis: Int64
-    let requestId: String
-
-    var id: String {
-        self.requestId
-    }
-
-    var notificationBodyText: String {
-        self.content.notificationBodyText
-    }
-
-    var cardId: String? {
-        self.content.cardId
-    }
-}
-
-enum ScheduledReviewNotificationPayloadContent: Hashable, Sendable {
-    case card(cardId: String, frontText: String)
-    case fallback
-
-    var cardId: String? {
-        switch self {
-        case .card(let cardId, _):
-            return cardId
-        case .fallback:
-            return nil
-        }
-    }
-
-    var notificationBodyText: String {
-        switch self {
-        case .card(_, let frontText):
-            return frontText
-        case .fallback:
-            return reviewNotificationFallbackBodyText
-        }
-    }
-}
-
-extension ScheduledReviewNotificationPayloadContent: Codable {
-    private enum CodingKeys: String, CodingKey {
-        case kind
-        case cardId
-        case frontText
-    }
-
-    private enum Kind: String, Codable {
-        case card
-        case fallback
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        let kind = try container.decode(Kind.self, forKey: .kind)
-        switch kind {
-        case .card:
-            let cardId = try container.decode(String.self, forKey: .cardId)
-            let frontText = try container.decode(String.self, forKey: .frontText)
-            self = .card(cardId: cardId, frontText: frontText)
-        case .fallback:
-            self = .fallback
-        }
-    }
-
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        switch self {
-        case .card(let cardId, let frontText):
-            try container.encode(Kind.card, forKey: .kind)
-            try container.encode(cardId, forKey: .cardId)
-            try container.encode(frontText, forKey: .frontText)
-        case .fallback:
-            try container.encode(Kind.fallback, forKey: .kind)
-        }
-    }
-}
-
-extension ScheduledReviewNotificationPayload: Codable {
-    private enum CodingKeys: String, CodingKey {
-        case workspaceId
-        case reviewFilter
-        case content
-        case scheduledAtMillis
-        case requestId
-        case cardId
-        case frontText
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.workspaceId = try container.decode(String.self, forKey: .workspaceId)
-        self.reviewFilter = try container.decode(PersistedReviewFilter.self, forKey: .reviewFilter)
-        self.scheduledAtMillis = try container.decode(Int64.self, forKey: .scheduledAtMillis)
-        self.requestId = try container.decode(String.self, forKey: .requestId)
-
-        if let content = try container.decodeIfPresent(ScheduledReviewNotificationPayloadContent.self, forKey: .content) {
-            self.content = content
-            return
-        }
-
-        let cardId = try container.decode(String.self, forKey: .cardId)
-        let frontText = try container.decode(String.self, forKey: .frontText)
-        self.content = .card(cardId: cardId, frontText: frontText)
-    }
-
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        try container.encode(self.workspaceId, forKey: .workspaceId)
-        try container.encode(self.reviewFilter, forKey: .reviewFilter)
-        try container.encode(self.content, forKey: .content)
-        try container.encode(self.scheduledAtMillis, forKey: .scheduledAtMillis)
-        try container.encode(self.requestId, forKey: .requestId)
-    }
-}
-
-struct CurrentReviewNotificationCard: Hashable, Sendable {
-    let reviewFilter: PersistedReviewFilter
-    let cardId: String
-    let frontText: String
-}
-
-enum AppNotificationTapType: String, Codable, Hashable, Sendable {
-    case reviewReminder
-    case strictReminder
-}
-
-enum AppNotificationTapSource: String, Codable, Hashable, Sendable {
-    case notificationResponse = "notification_response"
-    case uiTestEnvironment = "ui_test_environment"
-}
-
-struct AppNotificationTapFallback: Codable, Hashable, Sendable {
-    let stage: String
-    let reason: String
-    let notificationType: String?
-    let details: String?
-}
-
-enum AppNotificationTapRequest: Codable, Hashable, Sendable {
-    case openReviewReminder
-    case openStrictReminder
-    case fallback(AppNotificationTapFallback)
-}
-
-struct PendingAppNotificationTapEnvelope: Codable, Hashable, Sendable {
-    let schemaVersion: Int
-    let request: AppNotificationTapRequest
-    let receivedAtMillis: Int64
-    let source: AppNotificationTapSource
-}
-
-struct ReviewNotificationSchedulingSnapshot: Sendable {
-    let databaseURL: URL?
-    let workspaceId: String
-    let reviewFilter: ReviewFilter
-    let now: Date
-    let settings: ReviewNotificationsSettings
-    let lastActiveAt: Date?
-    let pendingRequestLimit: Int
-}
-
-struct ScheduledReviewNotificationLoadResult: Sendable {
-    let payloads: [ScheduledReviewNotificationPayload]
-}
-
-enum ReviewNotificationPermissionStatus: Hashable, Sendable {
-    case allowed
-    case notRequested
-    case blocked
-
-    var title: String {
-        switch self {
-        case .allowed:
-            return String(
-                localized: "review_notification_permission.allowed",
-                table: "Foundation",
-                comment: "Notifications permission title for allowed"
-            )
-        case .notRequested:
-            return String(
-                localized: "review_notification_permission.not_requested",
-                table: "Foundation",
-                comment: "Notifications permission title for not requested"
-            )
-        case .blocked:
-            return String(
-                localized: "review_notification_permission.blocked",
-                table: "Foundation",
-                comment: "Notifications permission title for blocked"
-            )
-        }
-    }
-
-    var actionTitle: String {
-        switch self {
-        case .allowed, .blocked:
-            return String(
-                localized: "shared.action.open_settings",
-                table: "Foundation",
-                comment: "Notifications permission action title to open Settings"
-            )
-        case .notRequested:
-            return String(
-                localized: "review_notification_permission.allow_notifications",
-                table: "Foundation",
-                comment: "Notifications permission action title to allow notifications"
-            )
-        }
-    }
-}
-
-func makeDefaultReviewNotificationsSettings() -> ReviewNotificationsSettings {
-    // Internal reminder categories default to enabled. Delivery is still gated by
-    // the current system notification permission state.
-    ReviewNotificationsSettings(
-        isEnabled: true,
-        selectedMode: .daily,
-        daily: DailyReviewNotificationsSettings(
-            hour: defaultDailyReminderHour,
-            minute: defaultDailyReminderMinute
-        ),
-        inactivity: InactivityReviewNotificationsSettings(
-            windowStartHour: defaultDailyReminderHour,
-            windowStartMinute: defaultDailyReminderMinute,
-            windowEndHour: defaultInactivityReminderWindowEndHour,
-            windowEndMinute: defaultInactivityReminderWindowEndMinute,
-            idleMinutes: 120
-        ),
-        showAppIconBadge: true
-    )
-}
-
-func makeDefaultNotificationPermissionPromptState() -> NotificationPermissionPromptState {
-    NotificationPermissionPromptState(
-        hasShownPrePrompt: false,
-        hasRequestedSystemPermission: false,
-        hasDismissedPrePrompt: false
-    )
-}
-
-func hasEnoughReviewHistoryForNotificationPrompt(reviewCount: Int) -> Bool {
-    reviewCount >= reviewNotificationPermissionPromptThreshold
-}
-
-func reviewNotificationPendingRequestsLimit(strictRemindersSettings: StrictRemindersSettings) -> Int {
-    guard strictRemindersSettings.isEnabled else {
-        return appNotificationPendingRequestsLimit
-    }
-
-    return max(0, appNotificationPendingRequestsLimit - strictReminderPendingRequestsLimit)
-}
-
-func makeReviewNotificationsSettingsUserDefaultsKey(workspaceId: String) -> String {
-    "\(reviewNotificationsSettingsUserDefaultsKeyPrefix)\(workspaceId)"
-}
-
-func makeScheduledReviewNotificationsUserDefaultsKey(workspaceId: String) -> String {
-    "\(reviewNotificationScheduledPayloadsUserDefaultsKeyPrefix)\(workspaceId)"
-}
-
-func loadReviewNotificationsSettings(
-    userDefaults: UserDefaults,
-    decoder: JSONDecoder,
-    workspaceId: String?
-) -> ReviewNotificationsSettings {
-    guard
-        let workspaceId,
-        let data = userDefaults.data(forKey: makeReviewNotificationsSettingsUserDefaultsKey(workspaceId: workspaceId))
-    else {
-        return makeDefaultReviewNotificationsSettings()
-    }
-
-    do {
-        return try decoder.decode(ReviewNotificationsSettings.self, from: data)
-    } catch {
-        userDefaults.removeObject(forKey: makeReviewNotificationsSettingsUserDefaultsKey(workspaceId: workspaceId))
-        return makeDefaultReviewNotificationsSettings()
-    }
-}
-
-func loadNotificationPermissionPromptState(
-    userDefaults: UserDefaults,
-    decoder: JSONDecoder
-) -> NotificationPermissionPromptState {
-    guard let data = userDefaults.data(forKey: reviewNotificationPromptStateUserDefaultsKey) else {
-        return makeDefaultNotificationPermissionPromptState()
-    }
-
-    do {
-        return try decoder.decode(NotificationPermissionPromptState.self, from: data)
-    } catch {
-        userDefaults.removeObject(forKey: reviewNotificationPromptStateUserDefaultsKey)
-        return makeDefaultNotificationPermissionPromptState()
-    }
-}
-
-func loadScheduledReviewNotifications(
-    userDefaults: UserDefaults,
-    decoder: JSONDecoder,
-    workspaceId: String?
-) -> [ScheduledReviewNotificationPayload] {
-    guard
-        let workspaceId,
-        let data = userDefaults.data(forKey: makeScheduledReviewNotificationsUserDefaultsKey(workspaceId: workspaceId))
-    else {
-        return []
-    }
-
-    do {
-        return try decoder.decode([ScheduledReviewNotificationPayload].self, from: data)
-    } catch {
-        userDefaults.removeObject(forKey: makeScheduledReviewNotificationsUserDefaultsKey(workspaceId: workspaceId))
-        return []
-    }
-}
-
-func makeReviewNotificationRequestIdentifier(workspaceId: String, kind: String, suffix: String) -> String {
-    "review-notification::\(workspaceId)::\(kind)::\(suffix)"
-}
-
-/// Review reminders are identified by the shared `review-notification::` identifier prefix.
-func isReviewNotificationRequestIdentifier(identifier: String) -> Bool {
-    identifier.hasPrefix("review-notification::")
-}
-
-/// Keeps only identifiers that belong to review reminders.
-func filterReviewNotificationRequestIdentifiers(identifiers: [String]) -> [String] {
-    identifiers.filter(isReviewNotificationRequestIdentifier)
-}
-
-func appNotificationPendingRequestBreakdown(
-    identifiers: [String]
-) -> AppNotificationPendingRequestBreakdown {
-    let reviewCount: Int = identifiers.filter(isReviewNotificationRequestIdentifier).count
-    let strictCount: Int = identifiers.filter(isStrictReminderRequestIdentifier).count
-    return AppNotificationPendingRequestBreakdown(
-        totalCount: identifiers.count,
-        reviewCount: reviewCount,
-        strictCount: strictCount,
-        otherCount: identifiers.count - reviewCount - strictCount
-    )
-}
-
-func notificationScheduledAtMillisRange(
-    scheduledAtMillisValues: [Int64]
-) -> NotificationScheduledAtMillisRange {
-    NotificationScheduledAtMillisRange(
-        firstScheduledAtMillis: scheduledAtMillisValues.min(),
-        lastScheduledAtMillis: scheduledAtMillisValues.max()
-    )
-}
-
-func reviewNotificationScheduledAtMillisRange(
-    payloads: [ScheduledReviewNotificationPayload]
-) -> NotificationScheduledAtMillisRange {
-    notificationScheduledAtMillisRange(
-        scheduledAtMillisValues: payloads.map(\.scheduledAtMillis)
-    )
-}
-
-func notificationSchedulingDelaySecondsRange(
-    scheduledAtMillisValues: [Int64],
-    now: Date
-) -> NotificationSchedulingDelaySecondsRange {
-    let delaySecondsValues: [Int] = scheduledAtMillisValues.map { scheduledAtMillis in
-        let rawDelaySeconds: TimeInterval = TimeInterval(scheduledAtMillis) / 1_000 - now.timeIntervalSince1970
-        return max(1, Int(rawDelaySeconds.rounded(.up)))
-    }
-
-    return NotificationSchedulingDelaySecondsRange(
-        minDelaySeconds: delaySecondsValues.min(),
-        maxDelaySeconds: delaySecondsValues.max()
-    )
-}
-
-func reviewNotificationSchedulingDelaySecondsRange(
-    payloads: [ScheduledReviewNotificationPayload],
-    now: Date
-) -> NotificationSchedulingDelaySecondsRange {
-    notificationSchedulingDelaySecondsRange(
-        scheduledAtMillisValues: payloads.map(\.scheduledAtMillis),
-        now: now
-    )
-}
-
-func reviewNotificationPermissionStatusDiagnosticValue(
-    status: ReviewNotificationPermissionStatus
-) -> String {
-    switch status {
-    case .allowed:
-        return "allowed"
-    case .notRequested:
-        return "not_requested"
-    case .blocked:
-        return "blocked"
-    }
-}
-
-func appNotificationApplicationStateDiagnosticValue(
-    applicationState: UIApplication.State
-) -> String {
-    switch applicationState {
-    case .active:
-        return "active"
-    case .inactive:
-        return "inactive"
-    case .background:
-        return "background"
-    @unknown default:
-        return "unknown"
-    }
-}
-
-@MainActor
-func currentAppNotificationApplicationStateDiagnosticValue() -> String {
-    appNotificationApplicationStateDiagnosticValue(
-        applicationState: UIApplication.shared.applicationState
-    )
-}
-
-func pendingAppNotificationRequestIdentifiers(
-    center: UNUserNotificationCenter
-) async -> [String] {
-    await withCheckedContinuation { continuation in
-        center.getPendingNotificationRequests { requests in
-            continuation.resume(returning: requests.map(\.identifier))
-        }
-    }
-}
-
-/// Returns the identifiers of pending review reminders queued by the app.
-func pendingReviewNotificationRequestIdentifiers(
-    center: UNUserNotificationCenter
-) async -> [String] {
-    await withCheckedContinuation { continuation in
-        center.getPendingNotificationRequests { requests in
-            continuation.resume(
-                returning: filterReviewNotificationRequestIdentifiers(
-                    identifiers: requests.map(\.identifier)
-                )
-            )
-        }
-    }
-}
-
-/// Returns the identifiers of delivered review reminders currently shown by Notification Center.
-func deliveredReviewNotificationRequestIdentifiers(
-    center: UNUserNotificationCenter
-) async -> [String] {
-    await withCheckedContinuation { continuation in
-        center.getDeliveredNotifications { notifications in
-            continuation.resume(
-                returning: filterReviewNotificationRequestIdentifiers(
-                    identifiers: notifications.map(\.request.identifier)
-                )
-            )
-        }
-    }
-}
-
-/// Removes delivered review reminders from Notification Center.
-func removeDeliveredReviewNotifications(
-    center: UNUserNotificationCenter
-) async {
-    let deliveredRequestIdentifiers = await deliveredReviewNotificationRequestIdentifiers(center: center)
-    guard deliveredRequestIdentifiers.isEmpty == false else {
-        return
-    }
-
-    center.removeDeliveredNotifications(withIdentifiers: deliveredRequestIdentifiers)
-}
-
-func makeReviewNotificationRequestSuffix(scheduledAt: Date, calendar: Calendar) -> String {
-    let dateFormatter = DateFormatter()
-    dateFormatter.calendar = calendar
-    dateFormatter.locale = Locale(identifier: "en_US_POSIX")
-    dateFormatter.dateFormat = "yyyy-MM-dd-HH-mm"
-    return dateFormatter.string(from: scheduledAt)
-}
-
-func makeReviewNotificationRequestIdentifiers(
-    workspaceId: String,
-    scheduledPayloads: [ScheduledReviewNotificationPayload]
-) -> [String] {
-    scheduledPayloads.map(\.requestId)
-}
-
-func acceptedReviewNotificationPayloads(
-    payloads: [ScheduledReviewNotificationPayload],
-    pendingRequestIdentifiers: [String]
-) -> [ScheduledReviewNotificationPayload] {
-    let pendingRequestIdentifierSet: Set<String> = Set(pendingRequestIdentifiers)
-    return payloads.filter { payload in
-        pendingRequestIdentifierSet.contains(payload.requestId)
-    }
-}
-
-func buildDailyReviewNotificationDates(
-    now: Date,
-    calendar: Calendar,
-    settings: DailyReviewNotificationsSettings
-) -> [Date] {
-    let startOfToday = calendar.startOfDay(for: now)
-
-    return (0..<dailyReminderSchedulingHorizonDays).compactMap { offset in
-        guard let day = calendar.date(byAdding: .day, value: offset, to: startOfToday) else {
-            return nil
-        }
-
-        let scheduledAt = calendar.date(
-            bySettingHour: settings.hour,
-            minute: settings.minute,
-            second: 0,
-            of: day
-        )
-        guard let scheduledAt, scheduledAt > now else {
-            return nil
-        }
-        return scheduledAt
-    }
-}
-
-func computeInactivityReminderDate(
-    settings: InactivityReviewNotificationsSettings,
-    lastActiveAt: Date,
-    calendar: Calendar
-) -> Date? {
-    guard settings.idleMinutes > 0 else {
-        return nil
-    }
-
-    let candidate = lastActiveAt.addingTimeInterval(TimeInterval(settings.idleMinutes * 60))
-    let candidateDay = calendar.startOfDay(for: candidate)
-    guard
-        let windowStart = calendar.date(
-            bySettingHour: settings.windowStartHour,
-            minute: settings.windowStartMinute,
-            second: 0,
-            of: candidateDay
-        ),
-        let windowEnd = calendar.date(
-            bySettingHour: settings.windowEndHour,
-            minute: settings.windowEndMinute,
-            second: 0,
-            of: candidateDay
-        ),
-        windowStart < windowEnd
-    else {
-        return nil
-    }
-
-    if candidate < windowStart {
-        return windowStart
-    }
-    if candidate <= windowEnd {
-        return candidate
-    }
-
-    guard let nextDay = calendar.date(byAdding: .day, value: 1, to: candidateDay) else {
-        return nil
-    }
-
-    return calendar.date(
-        bySettingHour: settings.windowStartHour,
-        minute: settings.windowStartMinute,
-        second: 0,
-        of: nextDay
-    )
-}
-
-func buildInactivityReviewNotificationDates(
-    lastActiveAt: Date,
-    now: Date,
-    calendar: Calendar,
-    settings: InactivityReviewNotificationsSettings
-) -> [Date] {
-    guard let firstScheduledAt = computeInactivityReminderDate(
-        settings: settings,
-        lastActiveAt: lastActiveAt,
-        calendar: calendar
-    ) else {
-        return []
-    }
-
-    let firstScheduledDay = calendar.startOfDay(for: firstScheduledAt)
-
-    return (0..<dailyReminderSchedulingHorizonDays).flatMap { offset -> [Date] in
-        let firstScheduledAtForDay: Date
-        if offset == 0 {
-            firstScheduledAtForDay = firstScheduledAt
-        } else {
-            guard
-                let day = calendar.date(byAdding: .day, value: offset, to: firstScheduledDay),
-                let nextScheduledAt = calendar.date(
-                    bySettingHour: settings.windowStartHour,
-                    minute: settings.windowStartMinute,
-                    second: 0,
-                    of: day
-                )
-            else {
-                return [Date]()
-            }
-            firstScheduledAtForDay = nextScheduledAt
-        }
-
-        return buildRepeatedInactivityReviewNotificationDatesForDay(
-            firstScheduledAt: firstScheduledAtForDay,
-            now: now,
-            calendar: calendar,
-            settings: settings
-        )
-    }
-}
-
-func buildRepeatedReviewNotificationPayloads(
-    workspaceId: String,
-    currentCard: CurrentReviewNotificationCard,
-    scheduledDates: [Date],
-    calendar: Calendar,
-    mode: ReviewNotificationMode
-) -> [ScheduledReviewNotificationPayload] {
-    return buildRepeatedReviewNotificationPayloads(
-        workspaceId: workspaceId,
-        reviewFilter: currentCard.reviewFilter,
-        content: .card(cardId: currentCard.cardId, frontText: currentCard.frontText),
-        scheduledDates: scheduledDates,
-        calendar: calendar,
-        mode: mode
-    )
-}
-
-func buildFallbackReviewNotificationPayloads(
-    workspaceId: String,
-    reviewFilter: PersistedReviewFilter,
-    scheduledDates: [Date],
-    calendar: Calendar,
-    mode: ReviewNotificationMode
-) -> [ScheduledReviewNotificationPayload] {
-    return buildRepeatedReviewNotificationPayloads(
-        workspaceId: workspaceId,
-        reviewFilter: reviewFilter,
-        content: .fallback,
-        scheduledDates: scheduledDates,
-        calendar: calendar,
-        mode: mode
-    )
-}
-
-private func buildRepeatedReviewNotificationPayloads(
-    workspaceId: String,
-    reviewFilter: PersistedReviewFilter,
-    content: ScheduledReviewNotificationPayloadContent,
-    scheduledDates: [Date],
-    calendar: Calendar,
-    mode: ReviewNotificationMode
-) -> [ScheduledReviewNotificationPayload] {
-    return scheduledDates.map { scheduledAt in
-        ScheduledReviewNotificationPayload(
-            workspaceId: workspaceId,
-            reviewFilter: reviewFilter,
-            content: content,
-            scheduledAtMillis: Int64(scheduledAt.timeIntervalSince1970 * 1000),
-            requestId: makeReviewNotificationRequestIdentifier(
-                workspaceId: workspaceId,
-                kind: mode.rawValue,
-                suffix: makeReviewNotificationRequestSuffix(
-                    scheduledAt: scheduledAt,
-                    calendar: calendar
-                )
-            )
-        )
-    }
-}
-
-func loadScheduledReviewNotificationPayloads(
-    snapshot: ReviewNotificationSchedulingSnapshot
-) async throws -> ScheduledReviewNotificationLoadResult {
-    guard let databaseURL = snapshot.databaseURL else {
-        return ScheduledReviewNotificationLoadResult(payloads: [])
-    }
-
-    return try await Task.detached(priority: .utility) {
-        let database = try LocalDatabase(databaseURL: databaseURL)
-        defer {
-            try? database.close()
-        }
-
-        let currentCard = try database.loadCurrentReviewNotificationCard(
-            workspaceId: snapshot.workspaceId,
-            reviewFilter: snapshot.reviewFilter,
-            now: snapshot.now
-        )
-
-        let calendar = Calendar.autoupdatingCurrent
-        let scheduledDates: [Date]
-        let mode = snapshot.settings.selectedMode
-        switch mode {
-        case .daily:
-            scheduledDates = buildDailyReviewNotificationDates(
-                now: snapshot.now,
-                calendar: calendar,
-                settings: snapshot.settings.daily
-            )
-        case .inactivity:
-            guard let lastActiveAt = snapshot.lastActiveAt else {
-                return ScheduledReviewNotificationLoadResult(payloads: [])
-            }
-            scheduledDates = buildInactivityReviewNotificationDates(
-                lastActiveAt: lastActiveAt,
-                now: snapshot.now,
-                calendar: calendar,
-                settings: snapshot.settings.inactivity
-            )
-        }
-
-        let limitedScheduledDates = Array(scheduledDates.prefix(max(0, snapshot.pendingRequestLimit)))
-        let payloads: [ScheduledReviewNotificationPayload]
-        if let currentCard {
-            payloads = buildRepeatedReviewNotificationPayloads(
-                workspaceId: snapshot.workspaceId,
-                currentCard: currentCard,
-                scheduledDates: limitedScheduledDates,
-                calendar: calendar,
-                mode: mode
-            )
-        } else {
-            payloads = buildFallbackReviewNotificationPayloads(
-                workspaceId: snapshot.workspaceId,
-                reviewFilter: makePersistedReviewFilter(reviewFilter: snapshot.reviewFilter),
-                scheduledDates: limitedScheduledDates,
-                calendar: calendar,
-                mode: mode
-            )
-        }
-
-        return ScheduledReviewNotificationLoadResult(payloads: payloads)
-    }.value
-}
-
-private func buildRepeatedInactivityReviewNotificationDatesForDay(
-    firstScheduledAt: Date,
-    now: Date,
-    calendar: Calendar,
-    settings: InactivityReviewNotificationsSettings
-) -> [Date] {
-    guard settings.idleMinutes > 0 else {
-        return []
-    }
-    guard
-        let windowEnd = calendar.date(
-            bySettingHour: settings.windowEndHour,
-            minute: settings.windowEndMinute,
-            second: 0,
-            of: firstScheduledAt
-        ),
-        firstScheduledAt <= windowEnd
-    else {
-        return []
-    }
-
-    var scheduledDates: [Date] = []
-    var nextScheduledAt: Date? = firstScheduledAt
-
-    while let currentScheduledAt = nextScheduledAt, currentScheduledAt <= windowEnd {
-        if currentScheduledAt > now {
-            scheduledDates.append(currentScheduledAt)
-        }
-
-        nextScheduledAt = calendar.date(
-            byAdding: .minute,
-            value: settings.idleMinutes,
-            to: currentScheduledAt
-        )
-    }
-
-    return scheduledDates
-}
-
-func buildAppNotificationUserInfo(notificationType: AppNotificationTapType) -> [AnyHashable: Any] {
-    return [
-        appNotificationTapTypeUserInfoKey: notificationType.rawValue
-    ]
-}
-
-func appNotificationTapType(request: AppNotificationTapRequest) -> String {
-    switch request {
-    case .openReviewReminder:
-        return AppNotificationTapType.reviewReminder.rawValue
-    case .openStrictReminder:
-        return AppNotificationTapType.strictReminder.rawValue
-    case .fallback(let fallback):
-        return fallback.notificationType ?? "fallback"
-    }
-}
-
-func makeDelayedNotificationSchedulingReadback(
-    pendingRequestIdentifiers: [String],
-    plannedRequestIdentifiers: [String]
-) -> DelayedNotificationSchedulingReadback {
-    let pendingRequestIdentifierSet: Set<String> = Set(pendingRequestIdentifiers)
-    return DelayedNotificationSchedulingReadback(
-        pending: appNotificationPendingRequestBreakdown(identifiers: pendingRequestIdentifiers),
-        recovered: plannedRequestIdentifiers.allSatisfy { requestIdentifier in
-            pendingRequestIdentifierSet.contains(requestIdentifier)
-        }
-    )
-}
-
-func delayedNotificationSchedulingReadback(
-    center: UNUserNotificationCenter,
-    plannedRequestIdentifiers: [String],
-    delayNanoseconds: UInt64
-) async throws -> DelayedNotificationSchedulingReadback {
-    try await Task.sleep(nanoseconds: delayNanoseconds)
-    let pendingRequestIdentifiers: [String] = await pendingAppNotificationRequestIdentifiers(center: center)
-    return makeDelayedNotificationSchedulingReadback(
-        pendingRequestIdentifiers: pendingRequestIdentifiers,
-        plannedRequestIdentifiers: plannedRequestIdentifiers
-    )
-}
-
-func makeNotificationSchedulingDiagnostics(
-    trigger: String,
-    scheduledAtMillisRange: NotificationScheduledAtMillisRange,
-    delaySecondsRange: NotificationSchedulingDelaySecondsRange,
-    pendingBeforeRequestIdentifiers: [String],
-    pendingAfterRequestIdentifiers: [String],
-    permissionStatusBefore: ReviewNotificationPermissionStatus,
-    permissionStatusAfter: ReviewNotificationPermissionStatus,
-    appStateBeforeAdd: String,
-    appStateAfterReadback: String,
-    delayedReadback: DelayedNotificationSchedulingReadback?
-) -> NotificationSchedulingDiagnostics {
-    NotificationSchedulingDiagnostics(
-        trigger: trigger,
-        pendingBefore: appNotificationPendingRequestBreakdown(
-            identifiers: pendingBeforeRequestIdentifiers
-        ),
-        pendingAfter: appNotificationPendingRequestBreakdown(
-            identifiers: pendingAfterRequestIdentifiers
-        ),
-        permissionStatusBefore: reviewNotificationPermissionStatusDiagnosticValue(
-            status: permissionStatusBefore
-        ),
-        permissionStatusAfter: reviewNotificationPermissionStatusDiagnosticValue(
-            status: permissionStatusAfter
-        ),
-        appStateBeforeAdd: appStateBeforeAdd,
-        appStateAfterReadback: appStateAfterReadback,
-        scheduledAtMillisRange: scheduledAtMillisRange,
-        delaySecondsRange: delaySecondsRange,
-        delayedReadback: delayedReadback
-    )
-}
-
-func makeNotificationSchedulingFailureWarning(
-    action: String,
-    scope: IOSObservationScope,
-    notificationKind: AppNotificationTapType,
-    workspaceId: String?,
-    requestId: String?,
-    stage: String,
-    plannedCount: Int,
-    acceptedCount: Int,
-    diagnostics: NotificationSchedulingDiagnostics,
-    error: Error?,
-    messageSummary: String?
-) -> NotificationSchedulingFailureWarning {
-    let nsError: NSError? = error.map { value in value as NSError }
-    let safeErrorDomain: String?
-    if let rawDomain = nsError?.domain {
-        let candidateDomain = safeDiagnosticIdentifier(rawDomain)
-        safeErrorDomain = candidateDomain == filteredDiagnosticValue ? nil : candidateDomain
-    } else {
-        safeErrorDomain = nil
-    }
-
-    return NotificationSchedulingFailureWarning(
-        action: action,
-        scope: scope,
-        notificationKind: notificationKind.rawValue,
-        workspaceId: workspaceId,
-        requestId: requestId,
-        stage: stage,
-        plannedCount: plannedCount,
-        acceptedCount: acceptedCount,
-        pendingBeforeCount: diagnostics.pendingBefore.totalCount,
-        pendingAfterCount: diagnostics.pendingAfter.totalCount,
-        errorDomain: safeErrorDomain,
-        errorCode: nsError?.code,
-        messageSummary: error.map { value in Flashcards.errorMessage(error: value) } ?? messageSummary,
-        diagnostics: diagnostics
-    )
-}
-
-func savePendingAppNotificationTap(
-    envelope: PendingAppNotificationTapEnvelope,
-    userDefaults: UserDefaults,
-    encoder: JSONEncoder
-) throws {
-    do {
-        let data = try encoder.encode(envelope)
-        userDefaults.set(data, forKey: pendingAppNotificationTapUserDefaultsKey)
-    } catch {
-        throw LocalStoreError.validation(
-            "Pending app notification tap could not be saved: \(Flashcards.errorMessage(error: error))"
-        )
-    }
-}
-
-func loadPendingAppNotificationTap(
-    userDefaults: UserDefaults,
-    decoder: JSONDecoder
-) throws -> PendingAppNotificationTapEnvelope? {
-    guard let data = userDefaults.data(forKey: pendingAppNotificationTapUserDefaultsKey) else {
-        return nil
-    }
-
-    do {
-        let envelope = try decoder.decode(PendingAppNotificationTapEnvelope.self, from: data)
-        guard envelope.schemaVersion == pendingAppNotificationTapSchemaVersion else {
-            throw LocalStoreError.validation(
-                "Pending app notification tap schema is unsupported: \(envelope.schemaVersion)"
-            )
-        }
-        return envelope
-    } catch {
-        throw LocalStoreError.validation(
-            "Pending app notification tap is invalid: \(Flashcards.errorMessage(error: error))"
-        )
-    }
-}
-
-func clearPendingAppNotificationTap(userDefaults: UserDefaults) {
-    userDefaults.removeObject(forKey: pendingAppNotificationTapUserDefaultsKey)
-}
-
-func logAppNotificationTapEvent(action: String, metadata: [String: String]) {
-    let observation = NotificationTapObservation(
-        action: NotificationTapAction(rawValue: action) ?? .fallback,
-        notificationType: metadata["notificationType"] ?? "unknown",
-        source: metadata["source"].flatMap(AppNotificationTapSource.init(rawValue:)),
-        appState: metadata["appState"],
-        scenePhaseAtConsume: metadata["scenePhaseAtConsume"],
-        receivedAtMillis: metadata["receivedAt"].flatMap(Int64.init),
-        stage: metadata["stage"]
-    )
-    if action == NotificationTapAction.dropped.rawValue || action == NotificationTapAction.fallback.rawValue {
-        FlashcardsObservability.captureWarning(
-            .notificationTapDropped(
-                NotificationTapDroppedWarning(
-                    observation: observation,
-                    reason: metadata["reason"] ?? "unspecified",
-                    detailSummary: metadata["details"]
-                )
-            )
-        )
-        return
-    }
-
-    FlashcardsObservability.addBreadcrumb(.notificationTap(observation))
-}
-
-func makeAppNotificationTapLogMetadata(
-    request: AppNotificationTapRequest,
-    source: AppNotificationTapSource?,
-    appState: String?,
-    scenePhase: String?,
-    receivedAtMillis: Int64?,
-    stage: String?,
-    reason: String?,
-    details: String?
-) -> [String: String] {
-    var metadata: [String: String] = [
-        "build": appBuildNumber(),
-        "notificationType": appNotificationTapType(request: request)
-    ]
-    if let source {
-        metadata["source"] = source.rawValue
-    }
-    if let appState {
-        metadata["appState"] = appState
-    }
-    if let scenePhase {
-        metadata["scenePhaseAtConsume"] = scenePhase
-    }
-    if let receivedAtMillis {
-        metadata["receivedAt"] = String(receivedAtMillis)
-    }
-    if let stage {
-        metadata["stage"] = stage
-    }
-    if let reason {
-        metadata["reason"] = reason
-    }
-    if let details {
-        metadata["details"] = details
-    }
-    return metadata
-}
-
-func parseAppNotificationTapRequest(userInfo: [AnyHashable: Any]) -> AppNotificationTapRequest? {
-    guard let rawNotificationType = userInfo[appNotificationTapTypeUserInfoKey] as? String else {
-        return nil
-    }
-    guard let notificationType = AppNotificationTapType(rawValue: rawNotificationType) else {
-        return .fallback(
-            AppNotificationTapFallback(
-                stage: "parse",
-                reason: "unsupported_notification_type",
-                notificationType: rawNotificationType,
-                details: nil
-            )
-        )
-    }
-
-    switch notificationType {
-    case .reviewReminder:
-        return .openReviewReminder
-    case .strictReminder:
-        return .openStrictReminder
-    }
-}
-
-func logAppNotificationTapFallback(fallback: AppNotificationTapFallback) {
-    let request = AppNotificationTapRequest.fallback(fallback)
-    let metadata = makeAppNotificationTapLogMetadata(
-        request: request,
-        source: nil,
-        appState: nil,
-        scenePhase: nil,
-        receivedAtMillis: nil,
-        stage: fallback.stage,
-        reason: fallback.reason,
-        details: fallback.details
-    )
-    logAppNotificationTapEvent(action: "notification_tap_fallback", metadata: metadata)
-}
-
-func reviewNotificationPermissionStatus(authorizationStatus: UNAuthorizationStatus) -> ReviewNotificationPermissionStatus {
-    switch authorizationStatus {
-    case .authorized, .provisional, .ephemeral:
-        return .allowed
-    case .notDetermined:
-        return .notRequested
-    case .denied:
-        return .blocked
-    @unknown default:
-        return .blocked
-    }
-}
-
-func resolveReviewNotificationPermissionStatus() async -> ReviewNotificationPermissionStatus {
-    let settings = await UNUserNotificationCenter.current().notificationSettings()
-    return reviewNotificationPermissionStatus(authorizationStatus: settings.authorizationStatus)
-}
